Refactoring function placements
